+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Correct the error message shown in the case of internal buildpack regex errors ([#76](https://github.com/heroku/procfile-cnb/pull/76)).
 - Updated `libcnb` and `libherokubuildpack` from 0.5.0 to 0.7.0 ([#49](https://github.com/heroku/procfile-cnb/pull/49) and [#60](https://github.com/heroku/procfile-cnb/pull/60)).
 
 ## 1.0.1